part1

+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: faguilar <faguilar@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/08/18 02:15:55 by faguilar          #+#    #+#             */
-/*   Updated: 2021/09/03 19:08:31 by faguilar         ###   ########.fr       */
+/*   Updated: 2021/09/03 20:33:56 by faguilar         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *s)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i] != '\0')
